ZCII-3811: fix job

+++ impl/src/java/ca/hec/jobs/impl/codeofconduct/ResetCodeOfConductJob.java
@@ -4,43 +4,41 @@ import javax.sql.DataSource;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.quartz.Job;
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.transaction.annotation.Transactional;
-import ca.hec.jobs.impl.AbstractQuartzJobImpl;
+import org.sakaiproject.db.cover.SqlService;
 
-public class ResetCodeOfConductJob extends AbstractQuartzJobImpl {
+public class ResetCodeOfConductJob implements Job {
 
 	private static Log log = LogFactory
 			.getLog(ResetCodeOfConductJob.class);
 
-	private JdbcTemplate jdbcTemplate;
-
-	public void setDataSource(DataSource dataSource) {
-		this.jdbcTemplate = new JdbcTemplate(dataSource);
-	}
-
+	private SqlService sqlService;
+	
 	@Transactional
 	@Override
 	public void execute(JobExecutionContext arg0) throws JobExecutionException {
-		int nbAcceptedCC = 0;
+		boolean success = false;
 
 		log.info("starting ResetCodeOfConductJob");
 	
-		loginToSakai("ResetCodeOfConductJob");
-
 		// copy code_of_conduct to code_of_conduct_delete
 		String backupDataSql = "insert into code_of_conduct_delete select * from code_of_conduct";
-		jdbcTemplate.execute(backupDataSql);
+		success = sqlService.dbWrite(backupDataSql);
 
-		// empty code_of_conduct table
-		String emptyTableSql = "delete from code_of_conduct";
-		nbAcceptedCC = jdbcTemplate.update(emptyTableSql);
+		if (success) {
+			// empty code_of_conduct table
+			String emptyTableSql = "delete from code_of_conduct";
+			success = sqlService.dbWrite(emptyTableSql);
+		}
 		
-    	logoutFromSakai();
-    	log.info("Moved " + nbAcceptedCC + " accepted code of conduct. ");
-
+		if (success) {
+		    	log.info("Successfully copied and deleted code of conduct data. ");
+		}
+		else {
+			log.error("Error removing code of conduct data.");
+		}
 	}
-
 }
